Mirroring: Preserve file permissions by uploading a .tar.xz as the build artifact (#18382)
GitHub's artifact mechanism does not preserve file permissions (or case,
but presumably if that were a problem we'd have known about it before
now).

The official workaround[1] is to use `tar` to collect your files and
upload that archive file as the artifact. So let's do that. As a bonus,
it might speed up the artifact upload and download.

[1]: https://github.com/actions/upload-artifact#maintaining-file-permissions-and-case-sensitive-files\n\nCommitted via a GitHub action: https://github.com/automattic/jetpack/runs/493888616